Split code generation into separate crate
Code generation is now done by the probe-rs-t2rust crate, which
makes it easier to run it separately to inspect the output.

The generated code is now also formatted with rustfmt, to improve
error messages.
