+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     "probe-rs",
+    "probe-rs-t2rust",
     "cli",
     "cargo-flash",
 ]
